feat: Add db connection functionality
A new file, db.go, has been added to the infra/db directory. This file includes the function to connect to the database, along with error handling for unsuccessful connections. It also loads the .env files, logs errors and enables debug mode based on environment variables. If AutoMigrateDb is set to true, DB migrations are executed for the Bank, Account, PixKey, and Transaction models.