feat: Add Pix key functionality to gRPC service
Added the PixKey.proto file to define gRPC services and messages for registration and retrieval of Pix keys, part of the Code_pix project. This is an essential step in providing Pix key functionality over gRPC. Also, added dependencies for protobuf in go.mod and go.sum to ensure successful protobuf compilation and use in the project.