feat: Move proto file and update package path
The pixkey.proto file was relocated to 'application/grpc/protofiles', and the protobuf package path was updated from github.com.juliofilizzola.Code_pix to github.com.juliofilizzola.codepix, and the go_package option was updated to "./protofiles;./pb". These changes were necessary due to a project restructuring and to ensure that gRPC services are able to locate the proto file and its associated package correctly.