feat: Add PixService gRPC generated code
This commit adds a newly generated file `pixkey_grpc.pb.go` which constitutes the gRPC client and server interfaces for the PixService. This is part of the Pixkeys feature being implemented to enable registration and retrieval of Pixkeys via a gRPC interface. It uses protoc-gen-go-grpc for auto-generation of this compliant code.