feat: Add gRPC Pix service to handle key registration
Introduced a new gRPC service, PixGrpcService, in the 'grpc' package to handle Pix key registration. This was needed to allow our application to receive and process Pix key registration requests via gRPC. This service uses the PixUseCase for registering the key and returning an appropriate response. It will improve the flexibility and efficiency of handling key registration requests.